automation/pyproject.toml
[tool.poetry]
name = "auto-activate-sm-workflows"
version = "1.0.0"
description = "Test Automation Workflow"
authors = ["HPE GLCP Team"]
packages = [
    { include = "scripts" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
allure-pytest = "^2.9.45"
requests = "^2.23.0"
allure-python-commons = "^2.8.22"
pytest-order = "^1.1.0"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
oauth2client = "^3.0.0"
google-api-core = "^1.23.0"
google-auth = "^1.23.0"
googleapis-common-protos = "^1.52.0"
pydantic = "1.10.4"
fastapi = "^0.59.0"
httpx = "^0.23.2"
boto3 = "^1.26.42"
hpe-glcp-automation-lib = "^2.399.0"
requests-oauthlib = "^1.3.0"
webdriver-manager = "^3.2.2"
names = "^0.3.0"
oauthlib = "^3.1.0"
bs4 = "^0.0.1"
lxml = "^4.6.1"
urllib3 = "^1.26.7"
okta = "^2.2.0"
pyotp = "^2.2.6"
humiolib = "^0.2.4"
launchdarkly-server-sdk = "^8.1.1"

[tool.poetry.dev-dependencies]
autoflake = "^1.3.1"
sqlalchemy-stubs = "^0.3"
flake8 = "3.9.2"
black = "22.3.0"
isort = "5.11.5"
requests-mock = "^1.8.0"
mock = "^4.0.2"
wily = "^1.19.0"
pylint = "^2.6.0"
fourmat = "^0.11.1"
pre-commit = "^3.3.3"

[[tool.poetry.source]]
name = "jfrog"
url = "https://aruba.jfrog.io/aruba/api/pypi/pypi-local/simple/"
secondary = true

[tool.poetry.scripts]
run = 'scripts.run:start'
lint = 'scripts.run:lint'
fix-lint = 'scripts.run:fix_lint'
test = 'scripts.run:test'
format-imports = 'scripts.run:format_imports'
cov = 'scripts.run:test_cov_html'
ci = 'scripts.run:run_ci'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 90

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
